ADM-358 remove next button and back button to configStep
